+++ features/Context/NavigationContext.php
@@ -10,6 +10,7 @@ use Pim\Bundle\CatalogBundle\Entity\AttributeGroup;
 use Pim\Bundle\CatalogBundle\Entity\Association;
 use Pim\Bundle\CatalogBundle\Entity\Category;
 use Pim\Bundle\CatalogBundle\Entity\Family;
+use Pim\Bundle\CatalogBundle\Entity\GroupType;
 
 /**
  * Context for navigating the website
@@ -56,6 +57,7 @@ class NavigationContext extends RawMinkContext implements PageObjectAwareInterfa
         'locales'                  => 'Locale index',
         'products'                 => 'Product index',
         'product groups'           => 'ProductGroup index',
+        'group types'              => 'GroupType index',
         'users'                    => 'User index',
         'user roles'               => 'UserRole index',
         'user groups'              => 'UserGroup index',
@@ -79,18 +81,9 @@ class NavigationContext extends RawMinkContext implements PageObjectAwareInterfa
      */
     public function resetCurrentPage()
     {
-        $this->getMainContext()->executeScript('sessionStorage.clear();');
+        $script = 'sessionStorage.clear(); typeof $ !== "undefined" && $(window).off("beforeunload");';
+        $this->getMainContext()->executeScript($script);
         $this->currentPage = null;
-    }
-
-    /**
-     * @AfterScenario
-     */
-    public function disableNavigationConfirmation()
-    {
-        if (strpos($this->currentPage, 'edit') || strpos($this->currentPage, 'creation')) {
-            $this->getMainContext()->executeScript('typeof $ !== "undefined" && $(window).off("beforeunload");');
-        }
     }
 
     /**
@@ -100,11 +93,8 @@ class NavigationContext extends RawMinkContext implements PageObjectAwareInterfa
      */
     public function iAmLoggedInAs($username)
     {
-        $password = $username;
-        $this->getFixturesContext()->getOrCreateUser($username, $password);
-
         $this->username = $username;
-        $this->password = $password;
+        $this->password = $username;
     }
 
     /**
@@ -227,8 +217,21 @@ class NavigationContext extends RawMinkContext implements PageObjectAwareInterfa
     public function iAmOnTheVariantGroupEditPage($identifier)
     {
         $page = 'VariantGroup';
-        $getter = sprintf('get%s', $page);
         $entity = $this->getFixturesContext()->getProductGroup($identifier);
+        $this->openPage(sprintf('%s edit', $page), array('id' => $entity->getId()));
+    }
+
+    /**
+     * @param string $identifier
+     *
+     * @Given /^I am on the "([^"]*)" group type page$/
+     * @Given /^I edit the "([^"]*)" group type$/
+     */
+    public function iAmOnTheGroupTypeEditPage($identifier)
+    {
+        $page = 'GroupType';
+        $getter = sprintf('get%s', $page);
+        $entity = $this->getFixturesContext()->$getter($identifier);
         $this->openPage(sprintf('%s edit', $page), array('id' => $entity->getId()));
     }
 
@@ -276,6 +279,17 @@ class NavigationContext extends RawMinkContext implements PageObjectAwareInterfa
     public function iShouldBeOnTheAttributeGroupPage(AttributeGroup $group)
     {
         $expectedAddress = $this->getPage('AttributeGroup edit')->getUrl(array('id' => $group->getId()));
+        $this->assertAddress($expectedAddress);
+    }
+
+    /**
+     * @param GroupType $groupType
+     *
+     * @Given /^I should be on the ("([^"]*)" group type) page$/
+     */
+    public function iShouldBeOnTheGroupTypePage(GroupType $groupType)
+    {
+        $expectedAddress = $this->getPage('GroupType edit')->getUrl(array('id' => $groupType->getId()));
         $this->assertAddress($expectedAddress);
     }