+++ src/Pim/Bundle/CatalogBundle/Tests/Unit/Form/Subscriber/IgnoreMissingFieldDataSubscriberTest.php
@@ -19,7 +19,7 @@ class IgnoreMissingFieldDataSubscriberTest extends \PHPUnit_Framework_TestCase
     /**
      * {@inheritdoc}
      */
-    public function setUp()
+    protected function setUp()
     {
         $this->subscriber = new IgnoreMissingFieldDataSubscriber();
     }