+++ src/Pim/Bundle/CatalogBundle/Tests/Unit/Form/Type/FamilyTypeTest.php
@@ -16,7 +16,7 @@ class FamilyTypeTest extends AbstractFormTypeTest
     /**
      * {@inheritdoc}
      */
-    public function setUp()
+    protected function setUp()
     {
         parent::setUp();