+++ src/Pim/Bundle/CatalogBundle/Tests/Unit/Validator/ConstraintGuesser/EmailGuesserTest.php
@@ -16,7 +16,7 @@ class EmailGuesserTest extends ConstraintGuesserTest
     /**
      * {@inheritdoc}
      */
-    public function setUp()
+    protected function setUp()
     {
         $this->target = new EmailGuesser();
     }