+++ src/Pim/Bundle/CatalogBundle/Tests/Unit/Validator/ConstraintGuesser/RangeGuesserTest.php
@@ -16,7 +16,7 @@ class RangeGuesserTest extends ConstraintGuesserTest
     /**
      * {@inheritdoc}
      */
-    public function setUp()
+    protected function setUp()
     {
         $this->target = new RangeGuesser();
     }