+++ src/Pim/Bundle/CatalogBundle/Tests/Unit/Validator/Constraints/FileValidatorTest.php
@@ -41,7 +41,7 @@ class FileValidatorTest extends \PHPUnit_Framework_TestCase
     /**
      * {@inheritdoc}
      */
-    public function setUp()
+    protected function setUp()
     {
         $this->target = new FileValidator();
     }