+++ src/Pim/Bundle/CatalogBundle/Tests/Unit/Validator/Constraints/ValidDefaultValueValidatorTest.php
@@ -21,7 +21,7 @@ class ValidDefaultValueValidatorTest extends \PHPUnit_Framework_TestCase
     /**
      * {@inheritdoc}
      */
-    public function setUp()
+    protected function setUp()
     {
         parent::setUp();
 
@@ -34,7 +34,7 @@ class ValidDefaultValueValidatorTest extends \PHPUnit_Framework_TestCase
     /**
      * {@inheritdoc}
      */
-    public function tearDown()
+    protected function tearDown()
     {
         $this->context = null;
         $this->validator = null;